+++ src/pages/Admin/Applicants/Filter/interfaces.ts
@@ -2,18 +2,9 @@ import { IUseApplicantsFilter } from "$hooks";
 import { ApplicantsFilter } from "$models/SearchFilters/ApplicantsFilter";
 import { ApplicantType } from "$interfaces/Applicant";
 import { ICareer } from "$interfaces/Career";
-import { FormikHelpers } from "formik";
 
-export interface ITranslations {
-  name: string;
-  submit: string;
-}
-
-interface ICommonProps {
+export interface IContainerProps {
   showFilter: boolean;
-}
-
-export interface IContainerProps extends ICommonProps {
   filter: ApplicantsFilter;
   refetchApplicants?: (filter: IUseApplicantsFilter) => void;
 }
@@ -23,11 +14,4 @@ export interface IFormValues {
   careers: ICareer[];
   applicantType: ApplicantType | "";
   _form: string;
-}
-
-export interface IComponentProps extends ICommonProps {
-  translations?: ITranslations;
-  initialValuesModel: ApplicantsFilter;
-  modelToValues: (filter?: ApplicantsFilter) => IFormValues;
-  onSubmit: (values: IFormValues, formikHelpers: FormikHelpers<IFormValues>) => Promise<any>;
 }
