+++ src/pages/Admin/components/Filter/component.tsx
@@ -1,24 +1,21 @@
-import React, { FunctionComponent } from "react";
+import React from "react";
 import classNames from "classnames";
-import { ApplicantType } from "$interfaces/Applicant";
 import { IComponentProps } from "./interfaces";
 
 import { FormFooter } from "$components/FormFooter";
 import { Formik } from "$components/Formik";
 import { FormikForm } from "$components/FormikForm";
-import { NameField } from "$components/Fields";
-import { CareerSelector } from "$components/CareerSelector";
-import { ApplicantTypeSelector } from "$components/ApplicantTypeSelector";
 
 import styles from "./styles.module.scss";
 
-export const Filter: FunctionComponent<IComponentProps> = ({
+export const Filter = <FormVariables,>({
   translations,
   initialValuesModel,
   modelToValues,
   onSubmit,
-  showFilter
-}) => (
+  showFilter,
+  children
+}: IComponentProps<FormVariables>) => (
   <Formik initialValues={modelToValues()} onSubmit={onSubmit}>
     {formikProps => (
       <FormikForm
@@ -29,20 +26,7 @@ export const Filter: FunctionComponent<IComponentProps> = ({
         modelToValues={modelToValues}
       >
         <div className={classNames(styles.fields, { [styles.hidden]: !showFilter })}>
-          {translations && (
-            <NameField
-              className={styles.name}
-              name="name"
-              label={translations.name}
-              withoutMargin
-            />
-          )}
-          <CareerSelector className={styles.careers} name="careers" />
-          <ApplicantTypeSelector
-            className={styles.applicantType}
-            name="applicantType"
-            excludedOptions={[ApplicantType.both]}
-          />
+          {children}
           <FormFooter
             className={styles.formFooter}
             isSubmitting={formikProps.isSubmitting}
