+++ src/pages/Admin/components/Filter/index.ts
@@ -1,0 +1,1 @@
+export { FilterContainer as Filter } from "./container";
